backend/pyproject.toml
[project]
name = "resume-enhancer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "emails>=0.6",
    "fastapi>=0.128.1",
    "jinja2>=3.1.6",
    "pwdlib[argon2,bcrypt]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "python-multipart>=0.0.22",
    "redis>=7.2.0",
    "sqlalchemy[asyncio]>=2.0.46",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "ruff>=0.15.2",
]

[tool.mypy]
python_version = "3.14"
strict = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = false
disallow_untyped_decorators = false
ignore_missing_imports = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_optional = true
exclude = "(migrations|.venv|uploads|email-templates)"


[tool.ruff]
line-length = 88
target-version = "py314"
exclude = ["migrations", ".venv"]


[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
    "T201",   # print statements are not allowed
    "ASYNC",
]
ignore = [
    "E501",  # line too long
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]


[tool.ruff.format]
quote-style = "single"
indent-style = "space"
